WAV Audio: Implement buffered reading and writing
The idea of this change is to replace single sample disc access
with buffered. This comes at a cost of extra stack memory usage
but significantly speeds up both reading and writing.

For reading this is entirely internal change: the reader simply
benefits from a pre-configured buffer (at compile time). The
buffer is used for both single and all sample readers. The buffer
size for all samples reader is bigger than for a single sample.

For writer a new API function has been added: WriteMultipleSamples
which takes a buffer of samples and bulk-writes them to the WAV
file.

It is possible to disable buffering by setting the following
compile definitions to 0:

- TL_AUDIO_WAV_READER_SAMPLE_FRAME_BUFFER_SIZE
- TL_AUDIO_WAV_READER_BUFFER_SIZE
- TL_AUDIO_WAV_WRITER_SAMPLE_FRAME_BUFFER_SIZE
- TL_AUDIO_WAV_WRITER_BUFFER_SIZE

Extra changes:
- Added Writer::GetFileSpec() to access configured specifications.
- Added [[nodiscard]] to internal functions to catch possible API
  mis-use more easily.
- Fixed incorrect check for the number of bytes written in the
  Writer.